updated BenchmarkExperiment
